otel-app/frontend/requirements.txt
Flask==3.0.3
fastapi==0.115.0
uvicorn==0.30.3
requests==2.32.3
opentelemetry-api==1.28.1
opentelemetry-sdk==1.28.1
opentelemetry-exporter-otlp==1.28.1
opentelemetry-instrumentation==0.49b1
opentelemetry-instrumentation-flask==0.49b1
opentelemetry-instrumentation-fastapi==0.49b1
opentelemetry-instrumentation-requests==0.49b1
